refactor: rename confirm-dialog overlay file
