+++ packages/confirm-dialog/src/vaadin-confirm-dialog-overlay.js
